third_party/WebKit/LayoutTests/virtual/threaded/fast/animationworklet/animation-worklet-animator-with-options.html

<!DOCTYPE html>

<script id="animate_with_options" type="text/worklet">
registerAnimator("test_animator", class {
  constructor(options) {
    console.log("animator constructor with options called.");
    console.log("options.dict.id = " + options.dict.id);
    console.log("options.dict.name = " + options.dict.name);
    console.log("options.string = " + options.string);
  }
  animate(currentTime, effect) {
    if (!this.animated)
        console.log("test animator animate called.");
    this.animated = true;
  }
});
</script>

<div id="target" style="will-change: transform, opacity;"></div>

<script src="resources/animation-worklet-tests.js"></script>
<script>
if (window.testRunner) {
  testRunner.waitUntilDone();
  testRunner.dumpAsText();
}

runInAnimationWorklet(
  document.getElementById('animate_with_options').textContent
).then(_ => {
  const effect = new KeyframeEffect(document.getElementById("target"), [{ opacity: 0 }], { duration: 1000 });
  const options = {
    dict: { 'name': 'test_name', 'id': 1 },
    string: 'test_string'
  };
  const animation = new WorkletAnimation('test_animator', [effect],
      document.timeline, options);
  animation.play();

  if (window.testRunner) {
    waitTwoAnimationFrames( _ => {
      testRunner.notifyDone();
    });
  }
});
</script>
